+++ test/com/onsoft/tiger/utils/AnnotatedMethodsMapperTest.ts
@@ -34,8 +34,16 @@ describe("AnnotatedMethodsMapper", ()=> {
       expect(mapper.getMethodByType(AnnotatedMethodType.BEFORE_ALL)).to.be.null;
     });
 
+    it("should return null when no AnnotatedMethodType.BEFORE_CLASS method is specified", function() {
+      expect(mapper.getMethodByType(AnnotatedMethodType.BEFORE_CLASS)).to.be.null;
+    });
+
     it("should return null when no AnnotatedMethodType.AFTER_ALL method is specified", function() {
       expect(mapper.getMethodByType(AnnotatedMethodType.AFTER_ALL)).to.be.null;
+    });
+
+    it("should return null when no AnnotatedMethodType.AFTER_CLASS method is specified", function() {
+      expect(mapper.getMethodByType(AnnotatedMethodType.AFTER_CLASS)).to.be.null;
     });
 
     it("should return null when no AnnotatedMethodType.BEFORE method is specified", function() {
@@ -54,8 +62,16 @@ describe("AnnotatedMethodsMapper", ()=> {
       expect(mapper.getMethodByType(AnnotatedMethodType.BEFORE_ALL)).to.equal(utils.BEFORE_ALL_METHOD);
     });
 
+    it("should return the registered method for the type AnnotatedMethodType.BEFORE_CLASS", function() {
+      expect(mapper.getMethodByType(AnnotatedMethodType.BEFORE_CLASS)).to.equal(utils.BEFORE_CLASS_METHOD);
+    });
+
     it("should return the registered method for the type AnnotatedMethodType.AFTER_ALL", function() {
       expect(mapper.getMethodByType(AnnotatedMethodType.AFTER_ALL)).to.equal(utils.AFTER_ALL_METHOD);
+    });
+
+    it("should return the registered method for the type AnnotatedMethodType.AFTER_CLASS", function() {
+      expect(mapper.getMethodByType(AnnotatedMethodType.AFTER_CLASS)).to.equal(utils.AFTER_CLASS_METHOD);
     });
 
     it("should return the registered method for the type AnnotatedMethodType.BEFORE", function() {