+++ utils/test-utils/classes/TestClassRunnerTestClass.ts
@@ -1,0 +1,17 @@
+export class TestClassRunnerTestClass {
+
+  public beforeAll():void {}
+
+  public static beforeClass():void {}
+
+  public before():void {}
+
+  public test():void {}
+  
+  public afterAll():void {}
+
+  public static afterClass():void {}
+
+  public after():void {}
+
+}