+++ utils/test-utils/utilities/DecoratorsTestUtils.ts
@@ -61,4 +61,5 @@ export const TEST_SUITE_PARAMS:TestSuiteParams = buildTestSuiteParams();
 export const DESCRIPTOR:PropertyDescriptor = null;
 export const KEY:string = "memberToTest";
 export const TARGET:any = {};
+export const STATIC_TARGET:Function = function(){};
 export const PARAMETER_INDEX:number = 0;